requirements.txt
requests
streamlit
python-dotenv
PyJWT
celery
pytest-mock
pytest
pytest-asyncio
sqlalchemy
whisperx
uvicorn
fastapi
python-multipart
openai-whisper
